move rms_helpers into a module and remove unneeded imports from comparison test
